+++ docs/conf.py
@@ -19,6 +19,7 @@
 #
 import os
 import sys
+import sphinx_bootstrap_theme
 from sphinx.domains.python import PythonDomain
 from sphinx.ext.autodoc import FunctionDocumenter
 sys.path.insert(0, os.path.abspath('.'))
@@ -90,13 +91,17 @@ todo_include_todos = False
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'nature'
+#html_theme = 'nature'
+html_theme = 'bootstrap'
+html_theme_path = sphinx_bootstrap_theme.get_html_theme_path()
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
 #
-# html_theme_options = {}
+html_theme_options = {
+    'bootswatch_theme': 'readable'
+}
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,